src/app/core/api/in-memory-data.service.ts
import { Injectable } from '@angular/core';
import { InMemoryDbService } from 'angular-in-memory-web-api';
import { Observable, of } from 'rxjs';
import { MessagesService } from 'src/app/share/component/display/messages/messages.service';
import { MOCK_HEROES } from '../mock/mock-heroes';
import { Hero } from '../model/hero';

@Injectable({
  providedIn: 'root',
})
export class InMemoryDataService implements InMemoryDbService {
  constructor(private messageService: MessagesService) {}

  createDb(): {} | Observable<{}> | Promise<{}> {
    const heroes = MOCK_HEROES;
    return { heroes };
  }

  genId(heroes: Hero[]): number {
    return heroes.length > 0
      ? Math.max(...heroes.map((hero) => hero.id)) + 1
      : 11;
  }

  handleError<T>(operation = 'operation', result?: T) {
    return (error: any): Observable<T> => {
      // Add error to messagesService
      this.messageService.add(`${operation} failed: ${error.message}`);

      // Let the app keep running by returning an empty result.
      return of(result as T);
    };
  }
}
